Fix Windows build

+++ src/appleseed.studio/mainwindow/mainwindow.cpp
@@ -323,7 +323,7 @@ QDockWidget* MainWindow::create_dock_widget(const char* dock_name)
 
     const auto& actions = m_ui->menu_view->actions();
     QAction* menu_separator = actions.last();
-    for (size_t i = static_cast<size_t>(actions.size()) - 2; i != 0; --i)
+    for (int i = actions.size() - 2; i != 0; --i)
         if (actions[i]->isSeparator())
         {
             menu_separator = actions[i];
